HotFix 1.2 display error fixed
