+++ src/components/authorization/QrScan/QrScan.css
@@ -1,7 +1,7 @@
 .qrScanner {
     position: fixed;
 
-    top: 209px;
+    top: 150px;
     left: 29px;
 
     width: 302px;
@@ -11,6 +11,7 @@
     display: flex;
     align-items: center;
     background-color: #D9D9D9;
+    overflow: hidden;
 
 
 }
@@ -18,34 +19,29 @@
 .textScan {
     position: absolute;
     color: #000;
-    left: 158px;
-    top: 103px;
+    left: 105px;
+    top: 85px;
     font-family: Alata, serif;
     font-size: 27px;
 }
 
-#qrCodeContainer{
-    position: relative;
-    width: 300px;
-    height: 300px;
+#qrCodeContainer {
+    position: absolute;
+    width: 100vw;
+    height: 100vh;
     border-radius: 7px;
     overflow: hidden;
-}
-
-#qr-shaded-region {
-    border-width: 50px;
 }
 
 .qr-message {
     position: absolute;
     top: 0;
     left: 0;
-    width: 100%;
-    height: 100%;
     background-color: #D9D9D9;
     color: black;
     font-size: 20px;
 }
+
 
 video {
     position: relative;
